requirements.txt
# Existing dependencies
fastapi
uvicorn
python-dotenv
requests
pydantic
sqlalchemy
psycopg2-binary
pytest
pytest-asyncio

# BranchSeeker/BizSight/LendSight/MergerMeter dependencies
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0
google-cloud-bigquery>=3.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
openpyxl>=3.0.0
reportlab>=3.6.0
Pillow>=10.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
anthropic>=0.7.0
openai>=1.0.0
numpy>=1.21.0
flask>=2.3.0
Werkzeug>=2.3.0
user-agents>=2.2.0
gunicorn>=21.2.0

# Optional - PDF Generation (for BizSight)
playwright>=1.40.0

# Optional - Production Server
# gunicorn>=20.1.0  # Already included above
# uwsgi>=2.0.20     # Alternative production server
census>=0.8.0
us>=3.0.0